add profile/logout to nav, logout actions
- index route conditionally displays lobby or login page based on auth
state (lobby route and lobby nav removed)
- cookie used to set auth state on page load
- nav shows profile and logout buttons when logged in
- reworked nav style, added hover style for buttons
