+++ packages/frontend/components/nav/navDesktop.tsx
@@ -1,17 +1,14 @@
 import React, { ReactElement } from 'react';
 import Link from 'next/link';
 
+import NavRight from './navRight';
+
 const NavDesktop = (): ReactElement => {
   return (
     <nav>
       <div className="nav_div">
-        <span className="before_nav_div" />
+        <div className="before_nav_div" />
         <ul>
-          <li>
-            <Link href="/lobby">
-              <a>Lobby</a>
-            </Link>
-          </li>
           <li>
             <Link href="/rules">
               <a>Rules</a>
@@ -33,7 +30,7 @@ const NavDesktop = (): ReactElement => {
             </Link>
           </li>
         </ul>
-        <span className="after_nav_div" />
+        <NavRight />
       </div>
 
       <style jsx>
@@ -42,63 +39,57 @@ const NavDesktop = (): ReactElement => {
             width: 100%;
           }
           nav > .nav_div {
-            border-top: 2px solid var(--gold);
-            border-bottom: 2px solid var(--gold);
+            display: flex;
             position: relative;
             width: 100%;
-            padding: 2px 50px;
+            padding: 0;
             margin: 0 auto;
           }
           nav > .nav_div > ul {
+            flex: 1;
             padding: 0;
             margin: 0;
             display: flex;
             justify-content: space-between;
-          }
-          // This is for the triangle border on the right side
-          nav > .nav_div > .after_nav_div {
-            position: absolute;
-            right: 0%;
-            top: -2px;
-            height: 41px;
-            width: 41px;
-            background-color: var(--gold);
-            clip-path: polygon(
-              100% 0%,
-              28% 50%,
-              100% 100%,
-              108% 100%,
-              35% 50%,
-              108% 0%
-            );
+            border-top: 2px solid var(--gold);
+            border-bottom: 2px solid var(--gold);
           }
           // Left side is identical to right side except for a vertical flip.
           nav > .nav_div > .before_nav_div {
-            position: absolute;
-            left: 0%;
-            top: -2px;
-            height: 41px;
-            width: 41px;
+            height: 36px;
+            width: 36px;
             background-color: var(--gold);
             clip-path: polygon(
-              100% 0%,
-              28% 50%,
+              0 0,
+              50% 50%,
+              0 100%,
               100% 100%,
-              108% 100%,
-              35% 50%,
-              108% 0%
+              100% 94.5%,
+              13% 94.5%,
+              58% 50%,
+              13% 5.5%,
+              100% 5.5%,
+              100% 0
             );
-            transform: rotate(180deg);
           }
           li {
             display: flex;
-            padding: 6px 8px;
+            flex-grow: 1;
+            list-style-type: none;
+            text-align: center;
           }
           a {
             color: var(--gold);
             text-decoration: none;
             font-size: 16px;
             font-family: Montserrat-Bold;
+            height: 100%;
+            width: 100%;
+            line-height: 32px;
+          }
+          a:hover {
+            background: var(--gold);
+            color: white;
           }
         `}
       </style>
