+++ packages/frontend/store/auth/action.types.ts
@@ -1,5 +1,7 @@
 export const LOGIN = 'LOGIN';
 export const SIGNUP = 'SIGNUP';
+export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
+export const LOGOUT = 'LOGOUT';
 
 // Redux Saga actions
 export interface ISignupAction {
@@ -14,3 +16,14 @@ export interface ILoginAction {
   username: string;
   password: string;
 }
+
+export interface ILoginSuccessAction {
+  type: typeof LOGIN_SUCCESS;
+  username: string;
+}
+
+export interface ILogoutAction {
+  type: typeof LOGOUT;
+}
+
+export type AuthActionTypes = ILoginSuccessAction | ILogoutAction;
