+++ packages/frontend/store/auth/actions.ts
@@ -1,4 +1,13 @@
-import { LOGIN, SIGNUP, ILoginAction, ISignupAction } from './action.types';
+import {
+  LOGIN,
+  SIGNUP,
+  LOGOUT,
+  ILoginAction,
+  ISignupAction,
+  ILogoutAction,
+  ILoginSuccessAction,
+  LOGIN_SUCCESS,
+} from './action.types';
 
 export const signup = (inputs: Omit<ISignupAction, 'type'>): ISignupAction => {
   return {
@@ -11,5 +20,20 @@ export const login = (inputs: Omit<ILoginAction, 'type'>): ILoginAction => {
   return {
     type: LOGIN,
     ...inputs,
+  };
+};
+
+export const loginSuccess = (
+  inputs: Omit<ILoginSuccessAction, 'type'>,
+): ILoginSuccessAction => {
+  return {
+    type: LOGIN_SUCCESS,
+    ...inputs,
+  };
+};
+
+export const logout = (): ILogoutAction => {
+  return {
+    type: LOGOUT,
   };
 };
